FO-2906: Endret navn til sendVedtak funksjonen
