+++ src/components/modal-controller.tsx
@@ -19,7 +19,7 @@ import {
 } from './modal/feil-modal/feil-modal-config';
 import { SpinnerModal } from './modal/spinner-modal/spinner-modal';
 import { VedtakSendtModal } from './modal/vedtak-sendt/vedtak-sendt-modal';
-import { BekreftVedtakSendtModal } from './modal/vedtak-sendt/bekreft-vedtak-sendt-modal';
+import { BekreftSendVedtakModal } from './modal/vedtak-sendt/bekreft-send-vedtak-modal';
 import SlettUtkastModal from './modal/slett-utkast-modal';
 import TaOverModal from './modal/ta-over-modal/ta-over-modal';
 
@@ -57,13 +57,13 @@ function finnFeilModalConfig(modalType: ModalType): FeilModalConfig | null {
 }
 
 export function ModalController() {
-	const { modalType } = useModalStore();
+	const { modalType, modalProps } = useModalStore();
 	const feilModalConfig = finnFeilModalConfig(modalType);
 
 	return (
 		<>
 			<SpinnerModal isOpen={modalType === ModalType.LASTER} />
-			<BekreftVedtakSendtModal isOpen={modalType === ModalType.BEKREFT_VEDTAK_SEND} />
+			<BekreftSendVedtakModal isOpen={modalType === ModalType.BEKREFT_SEND_VEDTAK} onSendVedtakBekreftet={modalProps.onSendVedtakBekreftet}/>
 			<VedtakSendtModal isOpen={modalType === ModalType.VEDTAK_SENT_SUKSESS} />
 			<SlettUtkastModal isOpen={modalType === ModalType.BEKREFT_SLETT_UTKAST} />
 			<TaOverModal isOpen={modalType === ModalType.BEKREFT_TA_OVER_UTKAST} />
