+++ src/pages/forhandsvisning/forhandsvisning.tsx
@@ -6,11 +6,13 @@ import env from '../../utils/environment';
 import { PILOT_TOGGLE, STOPPE_VEDTAKSUTSENDING_TOGGLE } from '../../rest/data/features';
 import { trengerBeslutter } from '../../components/utkast-skjema/skjema-utils';
 import { frontendlogger } from '../../utils/frontend-logger';
-import { lagHentForhandsvisningUrl } from '../../rest/api';
+import { useDataFetcherStore } from '../../stores/data-fetcher-store';
+import { lagHentForhandsvisningUrl, lagSendVedtakFetchInfo } from '../../rest/api';
+import { fetchWithInfo } from '../../rest/utils';
 import { useAppStore } from '../../stores/app-store';
 import { useViewStore, ViewType } from '../../stores/view-store';
 import { ModalType, useModalStore } from '../../stores/modal-store';
-import { useSkjemaStore } from '../../stores/skjema-store';
+import { useSkjemaStore} from '../../stores/skjema-store';
 import { erGodkjentAvBeslutter, finnUtkastAlltid } from '../../utils';
 import { getMockVedtaksbrevUrl } from '../../mock/mock-utils';
 import Show from '../../components/show';
@@ -21,9 +23,10 @@ import './forhandsvisning.less';
 export function Forhandsvisning() {
 	const { fnr } = useAppStore();
 	const { changeView } = useViewStore();
-	const { vedtak, features} = useDataStore();
+	const { vedtakFetcher } = useDataFetcherStore();
+	const { vedtak, features, oppfolgingData } = useDataStore();
 	const { showModal } = useModalStore();
-	const { innsatsgruppe } = useSkjemaStore();
+	const { innsatsgruppe, resetSkjema } = useSkjemaStore();
 	const { kanEndreUtkast } = useTilgangStore();
 
 	const [pdfStatus, setPdfStatus] = useState<PDFStatus>(PDFStatus.NOT_STARTED);
@@ -49,13 +52,31 @@ export function Forhandsvisning() {
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [pdfStatus]);
 
+
+	const sendVedtak = () => {
+		showModal(ModalType.LASTER);
+
+		fetchWithInfo(lagSendVedtakFetchInfo({ fnr }))
+			.then(() => {
+                resetSkjema();
+				vedtakFetcher.fetch({ fnr }, () => {
+					changeView(ViewType.HOVEDSIDE);
+					showModal(ModalType.VEDTAK_SENT_SUKSESS);
+				});
+			})
+			.catch(err => {
+				showModal(ModalType.FEIL_VED_SENDING);
+				frontendlogger.logMetrikk('feil-ved-sending', err);
+			});
+	};
+
 	const handleOnSendClicked = () => {
 		if (stoppeUtsendingfeatureToggle) {
 			showModal(ModalType.FEIL_UTSENDING_STOPPET);
 			return;
 		}
 
-		showModal(ModalType.BEKREFT_VEDTAK_SEND);
+		showModal(ModalType.BEKREFT_SEND_VEDTAK, { onSendVedtakBekreftet: sendVedtak });
 	};
 
 	return (
@@ -64,7 +85,12 @@ export function Forhandsvisning() {
 			<Footer className="forhandsvisning__footer">
 				<div className="forhandsvisning__aksjoner">
 					<Show if={kanEndreUtkast && erUtkastKlartTilUtsending}>
-						<Hovedknapp disabled={pdfStatus !== PDFStatus.SUCCESS} mini={true} onClick={handleOnSendClicked} className="forhandsvisning__knapp-sender">
+						<Hovedknapp
+							disabled={pdfStatus !== PDFStatus.SUCCESS}
+							mini={true}
+							onClick={handleOnSendClicked}
+							className="forhandsvisning__knapp-sender"
+						>
 							Send til bruker
 						</Hovedknapp>
 					</Show>
