+++ src/stores/modal-store.ts
@@ -3,7 +3,7 @@ import createUseContext from 'constate';
 
 export enum ModalType {
 	INGEN = 'INGEN',
-	BEKREFT_VEDTAK_SEND = 'BEKREFT_VEDTAK_SEND',
+	BEKREFT_SEND_VEDTAK = 'BEKREFT_SEND_VEDTAK',
 	VEDTAK_SENT_SUKSESS = 'VEDTAK_SENT_SUKSESS',
 	LASTER = 'LASTER_DATA',
 	BEKREFT_SLETT_UTKAST = 'BEKREFT_SLETT_UTKAST',
